Moving register base classes over to library and cleaning up structure.
