+++ src/107-Arduino-Cyphal.h
@@ -14,5 +14,7 @@
 
 #include "Node.h"
 #include "Types.h"
+#include "register/RegisterReadOnly.hpp"
+#include "register/RegisterReadWrite.hpp"
 
 #endif /* _107_ARDUINO_CYPHAL_H_ */
